@@ -1,20 +1,28 @@
 name: android_alarm_manager
 description: Flutter plugin for accessing the Android AlarmManager service, and
   running Dart code in the background when alarms fire.
-version: 0.4.1+6
-author: Flutter Team <flutter-dev@googlegroups.com>
+# 0.4.y+z is compatible with 1.0.0, if you land a breaking change bump
+# the version to 2.0.0.
+# See more details: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0
+version: 0.4.5+11
 homepage: https://github.com/flutter/plugins/tree/master/packages/android_alarm_manager
 
 dependencies:
   flutter:
     sdk: flutter
 
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  pedantic: ^1.8.0
+
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.androidalarmmanager
-    pluginClass: AndroidAlarmManagerPlugin
-    iosPrefix: FLT
+    platforms:
+      android:
+        package: io.flutter.plugins.androidalarmmanager
+        pluginClass: AndroidAlarmManagerPlugin
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.2.0 <2.0.0"
+  sdk: ">=2.1.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5 <2.0.0"
