@@ -1,21 +1,30 @@
 name: android_intent
 description: Flutter plugin for launching Android Intents. Not supported on iOS.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/flutter/plugins/tree/master/packages/android_intent
-version: 0.3.1
+# 0.3.y+z is compatible with 1.0.0, if you land a breaking change bump
+# the version to 2.0.0.
+# See more details: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0
+version: 0.3.7+3
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.androidintent
-    iosPrefix: FLT
-    pluginClass: AndroidIntentPlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.androidintent
+        pluginClass: AndroidIntentPlugin
 
 dependencies:
   flutter:
     sdk: flutter
-  platform: ^2.0.0
+  platform: ">=2.0.0 <4.0.0"
   meta: ^1.0.5
+dev_dependencies:
+  test: ^1.3.0
+  mockito: ^3.0.0
+  flutter_test:
+    sdk: flutter
+  pedantic: ^1.8.0
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.2.0 <2.0.0"
+  sdk: ">=2.3.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5 <2.0.0"
