@@ -2,13 +2,7 @@ name: camera
 description: A Flutter plugin for getting information about and controlling the
   camera on Android and iOS. Supports previewing the camera feed, capturing images, capturing video,
   and streaming image buffers to dart.
-version: 0.5.2+1
-authors:
-  - Flutter Team <flutter-dev@googlegroups.com>
-  - Luigi Agosti <luigi@tengio.com>
-  - Quentin Le Guennec <quentin@tengio.com>
-  - Koushik Ravikumar <koushik@tengio.com>
-  - Nissim Dsilva <nissim@tengio.com>
+version: 0.5.8+5
 
 homepage: https://github.com/flutter/plugins/tree/master/packages/camera
 
@@ -19,12 +13,21 @@ dependencies:
 dev_dependencies:
   path_provider: ^0.5.0
   video_player: ^0.10.0
+  flutter_test:
+    sdk: flutter
+  flutter_driver:
+    sdk: flutter
+  pedantic: ^1.8.0
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.camera
-    pluginClass: CameraPlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.camera
+        pluginClass: CameraPlugin
+      ios:
+        pluginClass: CameraPlugin
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.2.0 <2.0.0"
+  sdk: ">=2.1.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5 <2.0.0"
