@@ -14,10 +14,13 @@ dev_dependencies:
     sdk: flutter
   in_app_purchase:
     path: ../
+  integration_test:
+    path: ../../integration_test
+  pedantic: ^1.8.0
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=0.4.4 <2.0.0"
+  sdk: ">=2.3.0 <3.0.0"
+  flutter: ">=1.9.1+hotfix.2 <2.0.0"
