@@ -1,27 +1,37 @@
 name: package_info
 description: Flutter plugin for querying information about the application
   package, such as CFBundleVersion on iOS or versionCode on Android.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/flutter/plugins/tree/master/packages/package_info
-version: 0.4.0+4
+# 0.4.y+z is compatible with 1.0.0, if you land a breaking change bump
+# the version to 2.0.0.
+# See more details: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0
+version: 0.4.3
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.packageinfo
-    iosPrefix: FLT
-    pluginClass: PackageInfoPlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.packageinfo
+        pluginClass: PackageInfoPlugin
+      ios:
+        pluginClass: FLTPackageInfoPlugin
+      macos:
+        pluginClass: FLTPackageInfoPlugin
 
 dependencies:
   flutter:
     sdk: flutter
-  
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_driver:
     sdk: flutter
   test: any
+  integration_test:
+    path: ../integration_test
+  pedantic: ^1.8.0
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.5.0 <2.0.0"
+  sdk: ">=2.1.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5 <2.0.0"
