@@ -1,15 +1,20 @@
 name: sensors
 description: Flutter plugin for accessing the Android and iOS accelerometer and
   gyroscope sensors.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/flutter/plugins/tree/master/packages/sensors
-version: 0.4.0+2
+# 0.4.y+z is compatible with 1.0.0, if you land a breaking change bump
+# the version to 2.0.0.
+# See more details: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0
+version: 0.4.2+4
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.sensors
-    iosPrefix: FLT
-    pluginClass: SensorsPlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.sensors
+        pluginClass: SensorsPlugin
+      ios:
+        pluginClass: FLTSensorsPlugin
 
 dependencies:
   flutter:
@@ -19,7 +24,11 @@ dev_dependencies:
   test: ^1.3.0
   flutter_test:
     sdk: flutter
+  integration_test:
+    path: ../integration_test
+  mockito: ^4.1.1
+  pedantic: ^1.8.0
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=0.1.4 <2.0.0"
+  sdk: ">=2.1.0<3.0.0"
+  flutter: ">=1.12.13+hotfix.5 <2.0.0"
