@@ -1,18 +1,24 @@
 name: share
 description: Flutter plugin for sharing content via the platform share UI, using
   the ACTION_SEND intent on Android and UIActivityViewController on iOS.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/flutter/plugins/tree/master/packages/share
-version: 0.6.1+1
+# 0.6.y+z is compatible with 1.0.0, if you land a breaking change bump
+# the version to 2.0.0.
+# See more details: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0
+version: 0.6.5
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.share
-    iosPrefix: FLT
-    pluginClass: SharePlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.share
+        pluginClass: SharePlugin
+      ios:
+        pluginClass: FLTSharePlugin
 
 dependencies:
   meta: ^1.0.5
+  mime: ^0.9.7
   flutter:
     sdk: flutter
 
@@ -21,7 +27,10 @@ dev_dependencies:
   mockito: ^3.0.0
   flutter_test:
     sdk: flutter
+  integration_test:
+    path: ../integration_test
+  pedantic: ^1.8.0
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=0.1.4 <2.0.0"
+  sdk: ">=2.1.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5 <2.0.0"
