@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:ui' show hashValues;
 import 'package:flutter/foundation.dart';
 import 'package:collection/collection.dart';
 import 'package:json_annotation/json_annotation.dart';
@@ -55,6 +56,9 @@ class SkProductResponseWrapper {
         DeepCollectionEquality().equals(
             typedOther.invalidProductIdentifiers, invalidProductIdentifiers);
   }
+
+  @override
+  int get hashCode => hashValues(this.products, this.invalidProductIdentifiers);
 }
 
 /// Dart wrapper around StoreKit's [SKProductPeriodUnit](https://developer.apple.com/documentation/storekit/skproductperiodunit?language=objc).
@@ -110,6 +114,9 @@ class SKProductSubscriptionPeriodWrapper {
     final SKProductSubscriptionPeriodWrapper typedOther = other;
     return typedOther.numberOfUnits == numberOfUnits && typedOther.unit == unit;
   }
+
+  @override
+  int get hashCode => hashValues(this.numberOfUnits, this.unit);
 }
 
 /// Dart wrapper around StoreKit's [SKProductDiscountPaymentMode](https://developer.apple.com/documentation/storekit/skproductdiscountpaymentmode?language=objc).
@@ -187,6 +194,10 @@ class SKProductDiscountWrapper {
         typedOther.paymentMode == paymentMode &&
         typedOther.subscriptionPeriod == subscriptionPeriod;
   }
+
+  @override
+  int get hashCode => hashValues(this.price, this.priceLocale,
+      this.numberOfPeriods, this.paymentMode, this.subscriptionPeriod);
 }
 
 /// Dart wrapper around StoreKit's [SKProduct](https://developer.apple.com/documentation/storekit/skproduct?language=objc).
@@ -272,6 +283,17 @@ class SKProductWrapper {
         typedOther.subscriptionPeriod == subscriptionPeriod &&
         typedOther.introductoryPrice == introductoryPrice;
   }
+
+  @override
+  int get hashCode => hashValues(
+      this.productIdentifier,
+      this.localizedTitle,
+      this.localizedDescription,
+      this.priceLocale,
+      this.subscriptionGroupIdentifier,
+      this.price,
+      this.subscriptionPeriod,
+      this.introductoryPrice);
 }
 
 /// Object that indicates the locale of the price
@@ -282,7 +304,8 @@ class SKProductWrapper {
 //                 https://github.com/flutter/flutter/issues/26610
 @JsonSerializable()
 class SKPriceLocaleWrapper {
-  SKPriceLocaleWrapper({@required this.currencySymbol});
+  SKPriceLocaleWrapper(
+      {@required this.currencySymbol, @required this.currencyCode});
 
   /// Constructing an instance from a map from the Objective-C layer.
   ///
@@ -296,6 +319,9 @@ class SKPriceLocaleWrapper {
   ///The currency symbol for the locale, e.g. $ for US locale.
   final String currencySymbol;
 
+  ///The currency code for the locale, e.g. USD for US locale.
+  final String currencyCode;
+
   @override
   bool operator ==(Object other) {
     if (identical(other, this)) {
@@ -305,6 +331,10 @@ class SKPriceLocaleWrapper {
       return false;
     }
     final SKPriceLocaleWrapper typedOther = other;
-    return typedOther.currencySymbol == currencySymbol;
+    return typedOther.currencySymbol == currencySymbol &&
+        typedOther.currencyCode == currencyCode;
   }
+
+  @override
+  int get hashCode => hashValues(this.currencySymbol, this.currencyCode);
 }
