@@ -3,8 +3,8 @@ package io.flutter.plugins.androidalarmmanagerexample;
 import io.flutter.app.FlutterApplication;
 import io.flutter.plugin.common.PluginRegistry;
 import io.flutter.plugin.common.PluginRegistry.PluginRegistrantCallback;
-import io.flutter.plugins.GeneratedPluginRegistrant;
 import io.flutter.plugins.androidalarmmanager.AlarmService;
+import io.flutter.plugins.androidalarmmanager.AndroidAlarmManagerPlugin;
 
 public class Application extends FlutterApplication implements PluginRegistrantCallback {
   @Override
@@ -15,6 +15,7 @@ public class Application extends FlutterApplication implements PluginRegistrantC
 
   @Override
   public void registerWith(PluginRegistry registry) {
-    GeneratedPluginRegistrant.registerWith(registry);
+    AndroidAlarmManagerPlugin.registerWith(
+        registry.registrarFor("io.flutter.plugins.androidalarmmanager.AndroidAlarmManagerPlugin"));
   }
 }
