@@ -13,17 +13,16 @@ const String _backgroundName =
     'plugins.flutter.io/android_alarm_manager_background';
 
 // This is the entrypoint for the background isolate. Since we can only enter
-// an isolate once, we setup a MethodChannel to listen for method invokations
+// an isolate once, we setup a MethodChannel to listen for method invocations
 // from the native portion of the plugin. This allows for the plugin to perform
 // any necessary processing in Dart (e.g., populating a custom object) before
 // invoking the provided callback.
 void _alarmManagerCallbackDispatcher() {
-  const MethodChannel _channel =
-      MethodChannel(_backgroundName, JSONMethodCodec());
-
-  // Setup Flutter state needed for MethodChannels.
+  // Initialize state necessary for MethodChannels.
   WidgetsFlutterBinding.ensureInitialized();
 
+  const MethodChannel _channel =
+      MethodChannel(_backgroundName, JSONMethodCodec());
   // This is where the magic happens and we handle background events from the
   // native portion of the plugin.
   _channel.setMethodCallHandler((MethodCall call) async {
@@ -38,7 +37,15 @@ void _alarmManagerCallbackDispatcher() {
       print('Fatal: could not find callback');
       exit(-1);
     }
-    closure();
+
+    // ignore: inference_failure_on_function_return_type
+    if (closure is Function()) {
+      closure();
+      // ignore: inference_failure_on_function_return_type
+    } else if (closure is Function(int)) {
+      final int id = args[1];
+      closure(id);
+    }
   });
 
   // Once we've finished initializing, let the native portion of the plugin
@@ -46,14 +53,34 @@ void _alarmManagerCallbackDispatcher() {
   _channel.invokeMethod<void>('AlarmService.initialized');
 }
 
+// A lambda that returns the current instant in the form of a [DateTime].
+typedef DateTime _Now();
+// A lambda that gets the handle for the given [callback].
+typedef CallbackHandle _GetCallbackHandle(Function callback);
+
 /// A Flutter plugin for registering Dart callbacks with the Android
 /// AlarmManager service.
 ///
 /// See the example/ directory in this package for sample usage.
 class AndroidAlarmManager {
   static const String _channelName = 'plugins.flutter.io/android_alarm_manager';
-  static const MethodChannel _channel =
-      MethodChannel(_channelName, JSONMethodCodec());
+  static MethodChannel _channel =
+      const MethodChannel(_channelName, JSONMethodCodec());
+  // Function used to get the current time. It's [DateTime.now] by default.
+  static _Now _now = () => DateTime.now();
+  // Callback used to get the handle for a callback. It's
+  // [PluginUtilities.getCallbackHandle] by default.
+  static _GetCallbackHandle _getCallbackHandle =
+      (Function callback) => PluginUtilities.getCallbackHandle(callback);
+
+  /// This is exposed for the unit tests. It should not be accessed by users of
+  /// the plugin.
+  @visibleForTesting
+  static void setTestOverides(
+      {_Now now, _GetCallbackHandle getCallbackHandle}) {
+    _now = (now ?? _now);
+    _getCallbackHandle = (getCallbackHandle ?? _getCallbackHandle);
+  }
 
   /// Starts the [AndroidAlarmManager] service. This must be called before
   /// setting any alarms.
@@ -62,7 +89,7 @@ class AndroidAlarmManager {
   /// failure.
   static Future<bool> initialize() async {
     final CallbackHandle handle =
-        PluginUtilities.getCallbackHandle(_alarmManagerCallbackDispatcher);
+        _getCallbackHandle(_alarmManagerCallbackDispatcher);
     if (handle == null) {
       return false;
     }
@@ -80,12 +107,23 @@ class AndroidAlarmManager {
   /// `callback` must be either a top-level function or a static method from a
   /// class.
   ///
+  /// `callback` can be `Function()` or `Function(int)`
+  ///
   /// The timer is uniquely identified by `id`. Calling this function again
-  /// again with the same `id` will cancel and replace the existing timer.
+  /// with the same `id` will cancel and replace the existing timer.
+  ///
+  /// `id` will passed to `callback` if it is of type `Function(int)`
+  ///
+  /// If `alarmClock` is passed as `true`, the timer will be created with
+  /// Android's `AlarmManagerCompat.setAlarmClock`.
+  ///
+  /// If `allowWhileIdle` is passed as `true`, the timer will be created with
+  /// Android's `AlarmManagerCompat.setExactAndAllowWhileIdle` or
+  /// `AlarmManagerCompat.setAndAllowWhileIdle`.
   ///
   /// If `exact` is passed as `true`, the timer will be created with Android's
-  /// `AlarmManager.setRepeating`. When `exact` is `false` (the default), the
-  /// timer will be created with `AlarmManager.setInexactRepeating`.
+  /// `AlarmManagerCompat.setExact`. When `exact` is `false` (the default), the
+  /// timer will be created with `AlarmManager.set`.
   ///
   /// If `wakeup` is passed as `true`, the device will be woken up when the
   /// alarm fires. If `wakeup` is false (the default), the device will not be
@@ -100,22 +138,87 @@ class AndroidAlarmManager {
   static Future<bool> oneShot(
     Duration delay,
     int id,
-    dynamic Function() callback, {
+    Function callback, {
+    bool alarmClock = false,
+    bool allowWhileIdle = false,
+    bool exact = false,
+    bool wakeup = false,
+    bool rescheduleOnReboot = false,
+  }) =>
+      oneShotAt(
+        _now().add(delay),
+        id,
+        callback,
+        alarmClock: alarmClock,
+        allowWhileIdle: allowWhileIdle,
+        exact: exact,
+        wakeup: wakeup,
+        rescheduleOnReboot: rescheduleOnReboot,
+      );
+
+  /// Schedules a one-shot timer to run `callback` at `time`.
+  ///
+  /// The `callback` will run whether or not the main application is running or
+  /// in the foreground. It will run in the Isolate owned by the
+  /// AndroidAlarmManager service.
+  ///
+  /// `callback` must be either a top-level function or a static method from a
+  /// class.
+  ///
+  /// `callback` can be `Function()` or `Function(int)`
+  ///
+  /// The timer is uniquely identified by `id`. Calling this function again
+  /// with the same `id` will cancel and replace the existing timer.
+  ///
+  /// `id` will passed to `callback` if it is of type `Function(int)`
+  ///
+  /// If `alarmClock` is passed as `true`, the timer will be created with
+  /// Android's `AlarmManagerCompat.setAlarmClock`.
+  ///
+  /// If `allowWhileIdle` is passed as `true`, the timer will be created with
+  /// Android's `AlarmManagerCompat.setExactAndAllowWhileIdle` or
+  /// `AlarmManagerCompat.setAndAllowWhileIdle`.
+  ///
+  /// If `exact` is passed as `true`, the timer will be created with Android's
+  /// `AlarmManagerCompat.setExact`. When `exact` is `false` (the default), the
+  /// timer will be created with `AlarmManager.set`.
+  ///
+  /// If `wakeup` is passed as `true`, the device will be woken up when the
+  /// alarm fires. If `wakeup` is false (the default), the device will not be
+  /// woken up to service the alarm.
+  ///
+  /// If `rescheduleOnReboot` is passed as `true`, the alarm will be persisted
+  /// across reboots. If `rescheduleOnReboot` is false (the default), the alarm
+  /// will not be rescheduled after a reboot and will not be executed.
+  ///
+  /// Returns a [Future] that resolves to `true` on success and `false` on
+  /// failure.
+  static Future<bool> oneShotAt(
+    DateTime time,
+    int id,
+    Function callback, {
+    bool alarmClock = false,
+    bool allowWhileIdle = false,
     bool exact = false,
     bool wakeup = false,
     bool rescheduleOnReboot = false,
   }) async {
-    final int now = DateTime.now().millisecondsSinceEpoch;
-    final int first = now + delay.inMilliseconds;
-    final CallbackHandle handle = PluginUtilities.getCallbackHandle(callback);
+    // ignore: inference_failure_on_function_return_type
+    assert(callback is Function() || callback is Function(int));
+    assert(id.bitLength < 32);
+    final int startMillis = time.millisecondsSinceEpoch;
+    final CallbackHandle handle = _getCallbackHandle(callback);
     if (handle == null) {
       return false;
     }
-    final bool r = await _channel.invokeMethod<bool>('Alarm.oneShot', <dynamic>[
+    final bool r =
+        await _channel.invokeMethod<bool>('Alarm.oneShotAt', <dynamic>[
       id,
+      alarmClock,
+      allowWhileIdle,
       exact,
       wakeup,
-      first,
+      startMillis,
       rescheduleOnReboot,
       handle.toRawHandle(),
     ]);
@@ -131,8 +234,15 @@ class AndroidAlarmManager {
   /// `callback` must be either a top-level function or a static method from a
   /// class.
   ///
+  /// `callback` can be `Function()` or `Function(int)`
+  ///
   /// The repeating timer is uniquely identified by `id`. Calling this function
   /// again with the same `id` will cancel and replace the existing timer.
+  ///
+  /// `id` will passed to `callback` if it is of type `Function(int)`
+  ///
+  /// If `startAt` is passed, the timer will first go off at that time and
+  /// subsequently run with period `duration`.
   ///
   /// If `exact` is passed as `true`, the timer will be created with Android's
   /// `AlarmManager.setRepeating`. When `exact` is `false` (the default), the
@@ -151,15 +261,20 @@ class AndroidAlarmManager {
   static Future<bool> periodic(
     Duration duration,
     int id,
-    dynamic Function() callback, {
+    Function callback, {
+    DateTime startAt,
     bool exact = false,
     bool wakeup = false,
     bool rescheduleOnReboot = false,
   }) async {
-    final int now = DateTime.now().millisecondsSinceEpoch;
+    // ignore: inference_failure_on_function_return_type
+    assert(callback is Function() || callback is Function(int));
+    assert(id.bitLength < 32);
+    final int now = _now().millisecondsSinceEpoch;
     final int period = duration.inMilliseconds;
-    final int first = now + period;
-    final CallbackHandle handle = PluginUtilities.getCallbackHandle(callback);
+    final int first =
+        startAt != null ? startAt.millisecondsSinceEpoch : now + period;
+    final CallbackHandle handle = _getCallbackHandle(callback);
     if (handle == null) {
       return false;
     }
