@@ -1,154 +1,74 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
 package io.flutter.plugins.androidintent;
 
-import android.content.ComponentName;
-import android.content.Context;
-import android.content.Intent;
-import android.net.Uri;
-import android.os.Bundle;
-import android.provider.Settings;
-import android.util.Log;
-import io.flutter.plugin.common.MethodCall;
-import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
-import io.flutter.plugin.common.MethodChannel.Result;
+import androidx.annotation.NonNull;
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
-import java.util.ArrayList;
-import java.util.Map;
 
-/** AndroidIntentPlugin */
-@SuppressWarnings("unchecked")
-public class AndroidIntentPlugin implements MethodCallHandler {
-  private static final String TAG = AndroidIntentPlugin.class.getCanonicalName();
-  private final Registrar mRegistrar;
+/**
+ * Plugin implementation that uses the new {@code io.flutter.embedding} package.
+ *
+ * <p>Instantiate this in an add to app scenario to gracefully handle activity and context changes.
+ */
+public final class AndroidIntentPlugin implements FlutterPlugin, ActivityAware {
+  private final IntentSender sender;
+  private final MethodCallHandlerImpl impl;
 
-  /** Plugin registration. */
+  /**
+   * Initialize this within the {@code #configureFlutterEngine} of a Flutter activity or fragment.
+   *
+   * <p>See {@code io.flutter.plugins.androidintentexample.MainActivity} for an example.
+   */
+  public AndroidIntentPlugin() {
+    sender = new IntentSender(/*activity=*/ null, /*applicationContext=*/ null);
+    impl = new MethodCallHandlerImpl(sender);
+  }
+
+  /**
+   * Registers a plugin implementation that uses the stable {@code io.flutter.plugin.common}
+   * package.
+   *
+   * <p>Calling this automatically initializes the plugin. However plugins initialized this way
+   * won't react to changes in activity or context, unlike {@link AndroidIntentPlugin}.
+   */
   public static void registerWith(Registrar registrar) {
-    final MethodChannel channel =
-        new MethodChannel(registrar.messenger(), "plugins.flutter.io/android_intent");
-    channel.setMethodCallHandler(new AndroidIntentPlugin(registrar));
-  }
-
-  private AndroidIntentPlugin(Registrar registrar) {
-    this.mRegistrar = registrar;
-  }
-
-  private String convertAction(String action) {
-    switch (action) {
-      case "action_view":
-        return Intent.ACTION_VIEW;
-      case "action_voice":
-        return Intent.ACTION_VOICE_COMMAND;
-      case "settings":
-        return Settings.ACTION_SETTINGS;
-      default:
-        return action;
-    }
-  }
-
-  private Bundle convertArguments(Map<String, ?> arguments) {
-    Bundle bundle = new Bundle();
-    for (String key : arguments.keySet()) {
-      Object value = arguments.get(key);
-      if (value instanceof Integer) {
-        bundle.putInt(key, (Integer) value);
-      } else if (value instanceof String) {
-        bundle.putString(key, (String) value);
-      } else if (value instanceof Boolean) {
-        bundle.putBoolean(key, (Boolean) value);
-      } else if (value instanceof Double) {
-        bundle.putDouble(key, (Double) value);
-      } else if (value instanceof Long) {
-        bundle.putLong(key, (Long) value);
-      } else if (value instanceof byte[]) {
-        bundle.putByteArray(key, (byte[]) value);
-      } else if (value instanceof int[]) {
-        bundle.putIntArray(key, (int[]) value);
-      } else if (value instanceof long[]) {
-        bundle.putLongArray(key, (long[]) value);
-      } else if (value instanceof double[]) {
-        bundle.putDoubleArray(key, (double[]) value);
-      } else if (isTypedArrayList(value, Integer.class)) {
-        bundle.putIntegerArrayList(key, (ArrayList<Integer>) value);
-      } else if (isTypedArrayList(value, String.class)) {
-        bundle.putStringArrayList(key, (ArrayList<String>) value);
-      } else if (isStringKeyedMap(value)) {
-        bundle.putBundle(key, convertArguments((Map<String, ?>) value));
-      } else {
-        throw new UnsupportedOperationException("Unsupported type " + value);
-      }
-    }
-    return bundle;
-  }
-
-  private boolean isTypedArrayList(Object value, Class<?> type) {
-    if (!(value instanceof ArrayList)) {
-      return false;
-    }
-    ArrayList list = (ArrayList) value;
-    for (Object o : list) {
-      if (!(o == null || type.isInstance(o))) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  private boolean isStringKeyedMap(Object value) {
-    if (!(value instanceof Map)) {
-      return false;
-    }
-    Map map = (Map) value;
-    for (Object key : map.keySet()) {
-      if (!(key == null || key instanceof String)) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  private Context getActiveContext() {
-    return (mRegistrar.activity() != null) ? mRegistrar.activity() : mRegistrar.context();
+    IntentSender sender = new IntentSender(registrar.activity(), registrar.context());
+    MethodCallHandlerImpl impl = new MethodCallHandlerImpl(sender);
+    impl.startListening(registrar.messenger());
   }
 
   @Override
-  public void onMethodCall(MethodCall call, Result result) {
-    Context context = getActiveContext();
-    String action = convertAction((String) call.argument("action"));
+  public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
+    sender.setApplicationContext(binding.getApplicationContext());
+    sender.setActivity(null);
+    impl.startListening(binding.getBinaryMessenger());
+  }
 
-    // Build intent
-    Intent intent = new Intent(action);
-    if (mRegistrar.activity() == null) {
-      intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-    }
-    if (call.argument("category") != null) {
-      intent.addCategory((String) call.argument("category"));
-    }
-    if (call.argument("data") != null) {
-      intent.setData(Uri.parse((String) call.argument("data")));
-    }
-    if (call.argument("arguments") != null) {
-      intent.putExtras(convertArguments((Map) call.argument("arguments")));
-    }
-    if (call.argument("package") != null) {
-      String packageName = (String) call.argument("package");
-      intent.setPackage(packageName);
-      if (call.argument("componentName") != null) {
-        intent.setComponent(
-            new ComponentName(packageName, (String) call.argument("componentName")));
-      }
-      if (intent.resolveActivity(context.getPackageManager()) == null) {
-        Log.i(TAG, "Cannot resolve explicit intent - ignoring package");
-        intent.setPackage(null);
-      }
-    }
+  @Override
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+    sender.setApplicationContext(null);
+    sender.setActivity(null);
+    impl.stopListening();
+  }
 
-    Log.i(TAG, "Sending intent " + intent);
-    context.startActivity(intent);
+  @Override
+  public void onAttachedToActivity(@NonNull ActivityPluginBinding binding) {
+    sender.setActivity(binding.getActivity());
+  }
 
-    result.success(null);
+  @Override
+  public void onDetachedFromActivity() {
+    sender.setActivity(null);
+  }
+
+  @Override
+  public void onDetachedFromActivityForConfigChanges() {
+    onDetachedFromActivity();
+  }
+
+  @Override
+  public void onReattachedToActivityForConfigChanges(@NonNull ActivityPluginBinding binding) {
+    onAttachedToActivity(binding);
   }
 }
