@@ -14,27 +14,34 @@ _SerializedEnums _$_SerializedEnumsFromJson(Map json) {
 
 Map<String, dynamic> _$_SerializedEnumsToJson(_SerializedEnums instance) =>
     <String, dynamic>{
-      'response': _$SKPaymentTransactionStateWrapperEnumMap[instance.response]
+      'response': _$SKPaymentTransactionStateWrapperEnumMap[instance.response],
     };
 
-T _$enumDecode<T>(Map<T, dynamic> enumValues, dynamic source) {
+T _$enumDecode<T>(
+  Map<T, dynamic> enumValues,
+  dynamic source, {
+  T unknownValue,
+}) {
   if (source == null) {
     throw ArgumentError('A value must be provided. Supported values: '
         '${enumValues.values.join(', ')}');
   }
-  return enumValues.entries
-      .singleWhere((e) => e.value == source,
-          orElse: () => throw ArgumentError(
-              '`$source` is not one of the supported values: '
-              '${enumValues.values.join(', ')}'))
-      .key;
+
+  final value = enumValues.entries
+      .singleWhere((e) => e.value == source, orElse: () => null)
+      ?.key;
+
+  if (value == null && unknownValue == null) {
+    throw ArgumentError('`$source` is not one of the supported values: '
+        '${enumValues.values.join(', ')}');
+  }
+  return value ?? unknownValue;
 }
 
-const _$SKPaymentTransactionStateWrapperEnumMap =
-    <SKPaymentTransactionStateWrapper, dynamic>{
+const _$SKPaymentTransactionStateWrapperEnumMap = {
   SKPaymentTransactionStateWrapper.purchasing: 0,
   SKPaymentTransactionStateWrapper.purchased: 1,
   SKPaymentTransactionStateWrapper.failed: 2,
   SKPaymentTransactionStateWrapper.restored: 3,
-  SKPaymentTransactionStateWrapper.deferred: 4
+  SKPaymentTransactionStateWrapper.deferred: 4,
 };
