@@ -8,26 +8,28 @@ part of 'sk_payment_queue_wrapper.dart';
 
 SKError _$SKErrorFromJson(Map json) {
   return SKError(
-      code: json['code'] as int,
-      domain: json['domain'] as String,
-      userInfo: (json['userInfo'] as Map)?.map(
-        (k, e) => MapEntry(k as String, e),
-      ));
+    code: json['code'] as int,
+    domain: json['domain'] as String,
+    userInfo: (json['userInfo'] as Map)?.map(
+      (k, e) => MapEntry(k as String, e),
+    ),
+  );
 }
 
 Map<String, dynamic> _$SKErrorToJson(SKError instance) => <String, dynamic>{
       'code': instance.code,
       'domain': instance.domain,
-      'userInfo': instance.userInfo
+      'userInfo': instance.userInfo,
     };
 
 SKPaymentWrapper _$SKPaymentWrapperFromJson(Map json) {
   return SKPaymentWrapper(
-      productIdentifier: json['productIdentifier'] as String,
-      applicationUsername: json['applicationUsername'] as String,
-      requestData: json['requestData'] as String,
-      quantity: json['quantity'] as int,
-      simulatesAskToBuyInSandbox: json['simulatesAskToBuyInSandbox'] as bool);
+    productIdentifier: json['productIdentifier'] as String,
+    applicationUsername: json['applicationUsername'] as String,
+    requestData: json['requestData'] as String,
+    quantity: json['quantity'] as int,
+    simulatesAskToBuyInSandbox: json['simulatesAskToBuyInSandbox'] as bool,
+  );
 }
 
 Map<String, dynamic> _$SKPaymentWrapperToJson(SKPaymentWrapper instance) =>
@@ -36,5 +38,5 @@ Map<String, dynamic> _$SKPaymentWrapperToJson(SKPaymentWrapper instance) =>
       'applicationUsername': instance.applicationUsername,
       'requestData': instance.requestData,
       'quantity': instance.quantity,
-      'simulatesAskToBuyInSandbox': instance.simulatesAskToBuyInSandbox
+      'simulatesAskToBuyInSandbox': instance.simulatesAskToBuyInSandbox,
     };
