@@ -17,6 +17,33 @@ final PurchaseWrapper dummyPurchase = PurchaseWrapper(
   purchaseToken: 'purchaseToken',
   isAutoRenewing: false,
   originalJson: '',
+  developerPayload: 'dummy payload',
+  isAcknowledged: true,
+  purchaseState: PurchaseStateWrapper.purchased,
+);
+
+final PurchaseWrapper dummyUnacknowledgedPurchase = PurchaseWrapper(
+  orderId: 'orderId',
+  packageName: 'packageName',
+  purchaseTime: 0,
+  signature: 'signature',
+  sku: 'sku',
+  purchaseToken: 'purchaseToken',
+  isAutoRenewing: false,
+  originalJson: '',
+  developerPayload: 'dummy payload',
+  isAcknowledged: false,
+  purchaseState: PurchaseStateWrapper.purchased,
+);
+
+final PurchaseHistoryRecordWrapper dummyPurchaseHistoryRecord =
+    PurchaseHistoryRecordWrapper(
+  purchaseTime: 0,
+  signature: 'signature',
+  sku: 'sku',
+  purchaseToken: 'purchaseToken',
+  originalJson: '',
+  developerPayload: 'dummy payload',
 );
 
 void main() {
@@ -35,13 +62,25 @@ void main() {
       expect(details.purchaseID, dummyPurchase.orderId);
       expect(details.productID, dummyPurchase.sku);
       expect(details.transactionDate, dummyPurchase.purchaseTime.toString());
-      expect(details.verificationData.source, PurchaseSource.GooglePlay);
+      expect(details.verificationData.source, IAPSource.GooglePlay);
       expect(details.verificationData.localVerificationData,
           dummyPurchase.originalJson);
       expect(details.verificationData.serverVerificationData,
           dummyPurchase.purchaseToken);
       expect(details.skPaymentTransaction, null);
       expect(details.billingClientPurchase, dummyPurchase);
+      expect(details.pendingCompletePurchase, true);
+    });
+  });
+
+  group('PurchaseHistoryRecordWrapper', () {
+    test('converts from map', () {
+      final PurchaseHistoryRecordWrapper expected = dummyPurchaseHistoryRecord;
+      final PurchaseHistoryRecordWrapper parsed =
+          PurchaseHistoryRecordWrapper.fromJson(
+              buildPurchaseHistoryRecordMap(expected));
+
+      expect(parsed, equals(expected));
     });
   });
 
@@ -52,20 +91,53 @@ void main() {
         dummyPurchase,
         dummyPurchase
       ];
+      const String debugMessage = 'dummy Message';
+      final BillingResultWrapper billingResult = BillingResultWrapper(
+          responseCode: responseCode, debugMessage: debugMessage);
       final PurchasesResultWrapper expected = PurchasesResultWrapper(
-          responseCode: responseCode, purchasesList: purchases);
-
+          billingResult: billingResult,
+          responseCode: responseCode,
+          purchasesList: purchases);
       final PurchasesResultWrapper parsed =
           PurchasesResultWrapper.fromJson(<String, dynamic>{
+        'billingResult': buildBillingResultMap(billingResult),
         'responseCode': BillingResponseConverter().toJson(responseCode),
         'purchasesList': <Map<String, dynamic>>[
           buildPurchaseMap(dummyPurchase),
           buildPurchaseMap(dummyPurchase)
         ]
       });
-
+      expect(parsed.billingResult, equals(expected.billingResult));
       expect(parsed.responseCode, equals(expected.responseCode));
       expect(parsed.purchasesList, containsAll(expected.purchasesList));
+    });
+  });
+
+  group('PurchasesHistoryResult', () {
+    test('parsed from map', () {
+      final BillingResponse responseCode = BillingResponse.ok;
+      final List<PurchaseHistoryRecordWrapper> purchaseHistoryRecordList =
+          <PurchaseHistoryRecordWrapper>[
+        dummyPurchaseHistoryRecord,
+        dummyPurchaseHistoryRecord
+      ];
+      const String debugMessage = 'dummy Message';
+      final BillingResultWrapper billingResult = BillingResultWrapper(
+          responseCode: responseCode, debugMessage: debugMessage);
+      final PurchasesHistoryResult expected = PurchasesHistoryResult(
+          billingResult: billingResult,
+          purchaseHistoryRecordList: purchaseHistoryRecordList);
+      final PurchasesHistoryResult parsed =
+          PurchasesHistoryResult.fromJson(<String, dynamic>{
+        'billingResult': buildBillingResultMap(billingResult),
+        'purchaseHistoryRecordList': <Map<String, dynamic>>[
+          buildPurchaseHistoryRecordMap(dummyPurchaseHistoryRecord),
+          buildPurchaseHistoryRecordMap(dummyPurchaseHistoryRecord)
+        ]
+      });
+      expect(parsed.billingResult, equals(billingResult));
+      expect(parsed.purchaseHistoryRecordList,
+          containsAll(expected.purchaseHistoryRecordList));
     });
   });
 }
@@ -80,5 +152,27 @@ Map<String, dynamic> buildPurchaseMap(PurchaseWrapper original) {
     'purchaseToken': original.purchaseToken,
     'isAutoRenewing': original.isAutoRenewing,
     'originalJson': original.originalJson,
+    'developerPayload': original.developerPayload,
+    'purchaseState': PurchaseStateConverter().toJson(original.purchaseState),
+    'isAcknowledged': original.isAcknowledged,
+  };
+}
+
+Map<String, dynamic> buildPurchaseHistoryRecordMap(
+    PurchaseHistoryRecordWrapper original) {
+  return <String, dynamic>{
+    'purchaseTime': original.purchaseTime,
+    'signature': original.signature,
+    'sku': original.sku,
+    'purchaseToken': original.purchaseToken,
+    'originalJson': original.originalJson,
+    'developerPayload': original.developerPayload,
+  };
+}
+
+Map<String, dynamic> buildBillingResultMap(BillingResultWrapper original) {
+  return <String, dynamic>{
+    'responseCode': BillingResponseConverter().toJson(original.responseCode),
+    'debugMessage': original.debugMessage,
   };
 }
