@@ -6,6 +6,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart' show TestWidgetsFlutterBinding;
 import 'package:in_app_purchase/src/in_app_purchase/purchase_details.dart';
 import 'package:test/test.dart';
 
@@ -17,6 +18,8 @@ import 'package:in_app_purchase/store_kit_wrappers.dart';
 import '../store_kit_wrappers/sk_test_stub_objects.dart';
 
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
   final FakeIOSPlatform fakeIOSPlatform = FakeIOSPlatform();
 
   setUpAll(() {
@@ -58,7 +61,7 @@ void main() {
           .queryProductDetails(<String>['123', '456', '789'].toSet());
       expect(response.productDetails, []);
       expect(response.notFoundIDs, ['123', '456', '789']);
-      expect(response.error.source, PurchaseSource.AppStore);
+      expect(response.error.source, IAPSource.AppStore);
       expect(response.error.code, 'error_code');
       expect(response.error.message, 'error_message');
       expect(response.error.details, {'info': 'error_info'});
@@ -87,6 +90,28 @@ void main() {
       expect(response.error, isNull);
     });
 
+    test('queryPastPurchases should not block transaction updates', () async {
+      fakeIOSPlatform.transactions
+          .add(fakeIOSPlatform.createPurchasedTransactionWithProductID('foo'));
+      Completer completer = Completer();
+      Stream<List<PurchaseDetails>> stream =
+          AppStoreConnection.instance.purchaseUpdatedStream;
+
+      StreamSubscription subscription;
+      subscription = stream.listen((purchaseDetailsList) {
+        if (purchaseDetailsList.first.status == PurchaseStatus.purchased) {
+          completer.complete(purchaseDetailsList);
+          subscription.cancel();
+        }
+      });
+      QueryPurchaseDetailsResponse response =
+          await AppStoreConnection.instance.queryPastPurchases();
+      List<PurchaseDetails> result = await completer.future;
+      expect(result.length, 1);
+      expect(result.first.productID, 'foo');
+      expect(response.error, isNull);
+    });
+
     test('should get empty result if there is no restored transactions',
         () async {
       fakeIOSPlatform.testRestoredTransactionsNull = true;
@@ -105,7 +130,7 @@ void main() {
       QueryPurchaseDetailsResponse response =
           await AppStoreConnection.instance.queryPastPurchases();
       expect(response.pastPurchases, isEmpty);
-      expect(response.error.source, PurchaseSource.AppStore);
+      expect(response.error.source, IAPSource.AppStore);
       expect(response.error.message, 'error_test');
       expect(response.error.details, {'message': 'errorMessage'});
     });
@@ -128,7 +153,7 @@ void main() {
     test('should refresh receipt data', () async {
       PurchaseVerificationData receiptData =
           await AppStoreConnection.instance.refreshPurchaseVerificationData();
-      expect(receiptData.source, PurchaseSource.AppStore);
+      expect(receiptData.source, IAPSource.AppStore);
       expect(receiptData.localVerificationData, 'refreshed receipt data');
       expect(receiptData.serverVerificationData, 'refreshed receipt data');
     });
@@ -203,7 +228,7 @@ void main() {
       fakeIOSPlatform.testTransactionFail = true;
       List<PurchaseDetails> details = [];
       Completer completer = Completer();
-      PurchaseError error;
+      IAPError error;
 
       Stream<List<PurchaseDetails>> stream =
           AppStoreConnection.instance.purchaseUpdatedStream;
@@ -224,9 +249,9 @@ void main() {
       await AppStoreConnection.instance
           .buyNonConsumable(purchaseParam: purchaseParam);
 
-      PurchaseError completerError = await completer.future;
-      expect(completerError.code, kPurchaseErrorCode);
-      expect(completerError.source, PurchaseSource.AppStore);
+      IAPError completerError = await completer.future;
+      expect(completerError.code, 'purchase_error');
+      expect(completerError.source, IAPSource.AppStore);
       expect(completerError.message, 'ios_domain');
       expect(completerError.details, {'message': 'an error message'});
     });
@@ -242,7 +267,7 @@ void main() {
       subscription = stream.listen((purchaseDetailsList) {
         details.addAll(purchaseDetailsList);
         purchaseDetailsList.forEach((purchaseDetails) {
-          if (purchaseDetails.status == PurchaseStatus.purchased) {
+          if (purchaseDetails.pendingCompletePurchase) {
             AppStoreConnection.instance.completePurchase(purchaseDetails);
             completer.complete(details);
             subscription.cancel();
@@ -325,10 +350,10 @@ class FakeIOSPlatform {
 
   SKPaymentTransactionWrapper createPendingTransactionWithProductID(String id) {
     return SKPaymentTransactionWrapper(
+        transactionIdentifier: null,
         payment: SKPaymentWrapper(productIdentifier: id),
         transactionState: SKPaymentTransactionStateWrapper.purchasing,
         transactionTimeStamp: 123123.121,
-        transactionIdentifier: id,
         error: null,
         originalTransaction: null);
   }
@@ -346,10 +371,10 @@ class FakeIOSPlatform {
 
   SKPaymentTransactionWrapper createFailedTransactionWithProductID(String id) {
     return SKPaymentTransactionWrapper(
+        transactionIdentifier: null,
         payment: SKPaymentWrapper(productIdentifier: id),
         transactionState: SKPaymentTransactionStateWrapper.failed,
         transactionTimeStamp: 123123.121,
-        transactionIdentifier: id,
         error: SKError(
             code: 0,
             domain: 'ios_domain',
