@@ -136,9 +136,10 @@ void main() {
           details.transactionDate);
       expect(details.verificationData.localVerificationData, 'receipt data');
       expect(details.verificationData.serverVerificationData, 'receipt data');
-      expect(details.verificationData.source, PurchaseSource.AppStore);
+      expect(details.verificationData.source, IAPSource.AppStore);
       expect(details.skPaymentTransaction, dummyTransaction);
       expect(details.billingClientPurchase, null);
+      expect(details.pendingCompletePurchase, true);
     });
     test('Should generate correct map of the payment object', () {
       Map map = dummyPayment.toMap();
