@@ -22,6 +22,7 @@ final SKPaymentTransactionWrapper dummyOriginalTransaction =
   transactionIdentifier: '123123',
   error: dummyError,
 );
+
 final SKPaymentTransactionWrapper dummyTransaction =
     SKPaymentTransactionWrapper(
   transactionState: SKPaymentTransactionStateWrapper.purchased,
@@ -33,7 +34,7 @@ final SKPaymentTransactionWrapper dummyTransaction =
 );
 
 final SKPriceLocaleWrapper dummyLocale =
-    SKPriceLocaleWrapper(currencySymbol: '\$');
+    SKPriceLocaleWrapper(currencySymbol: '\$', currencyCode: 'USD');
 
 final SKProductSubscriptionPeriodWrapper dummySubscription =
     SKProductSubscriptionPeriodWrapper(
@@ -67,7 +68,10 @@ final SkProductResponseWrapper dummyProductResponseWrapper =
 );
 
 Map<String, dynamic> buildLocaleMap(SKPriceLocaleWrapper local) {
-  return {'currencySymbol': local.currencySymbol};
+  return {
+    'currencySymbol': local.currencySymbol,
+    'currencyCode': local.currencyCode
+  };
 }
 
 Map<String, dynamic> buildSubscriptionPeriodMap(
