@@ -4,38 +4,21 @@
 
 package io.flutter.plugins.quickactions;
 
-import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
-import android.content.pm.ShortcutInfo;
-import android.content.pm.ShortcutManager;
-import android.graphics.drawable.Icon;
-import android.os.Build;
-import android.os.Bundle;
-import io.flutter.plugin.common.MethodCall;
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
+import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
-import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 
 /** QuickActionsPlugin */
-@SuppressWarnings("unchecked")
-public class QuickActionsPlugin implements MethodCallHandler {
-  private final Registrar registrar;
+public class QuickActionsPlugin implements FlutterPlugin, ActivityAware {
+  private static final String CHANNEL_ID = "plugins.flutter.io/quick_actions";
 
-  // Channel is a static field because it needs to be accessible to the
-  // {@link ShortcutHandlerActivity} which has to be a static class with
-  // no-args constructor.
-  // It is also mutable because it is derived from {@link Registrar}.
-  private static MethodChannel channel;
-
-  private QuickActionsPlugin(Registrar registrar) {
-    this.registrar = registrar;
-  }
+  private MethodChannel channel;
+  private MethodCallHandlerImpl handler;
 
   /**
    * Plugin registration.
@@ -43,82 +26,49 @@ public class QuickActionsPlugin implements MethodCallHandler {
    * <p>Must be called when the application is created.
    */
   public static void registerWith(Registrar registrar) {
-    channel = new MethodChannel(registrar.messenger(), "plugins.flutter.io/quick_actions");
-    channel.setMethodCallHandler(new QuickActionsPlugin(registrar));
+    final QuickActionsPlugin plugin = new QuickActionsPlugin();
+    plugin.setupChannel(registrar.messenger(), registrar.context(), registrar.activity());
   }
 
   @Override
-  @SuppressLint("NewApi")
-  public void onMethodCall(MethodCall call, Result result) {
-    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N_MR1) {
-      // We already know that this functionality does not work for anything
-      // lower than API 25 so we chose not to return error. Instead we do nothing.
-      result.success(null);
-      return;
-    }
-    Context context = registrar.context();
-    ShortcutManager shortcutManager =
-        (ShortcutManager) context.getSystemService(Context.SHORTCUT_SERVICE);
-    switch (call.method) {
-      case "setShortcutItems":
-        List<Map<String, String>> serializedShortcuts = call.arguments();
-        List<ShortcutInfo> shortcuts = deserializeShortcuts(serializedShortcuts);
-        shortcutManager.setDynamicShortcuts(shortcuts);
-        break;
-      case "clearShortcutItems":
-        shortcutManager.removeAllDynamicShortcuts();
-        break;
-      default:
-        result.notImplemented();
-        return;
-    }
-    result.success(null);
+  public void onAttachedToEngine(FlutterPluginBinding binding) {
+    setupChannel(binding.getBinaryMessenger(), binding.getApplicationContext(), null);
   }
 
-  @SuppressLint("NewApi")
-  private List<ShortcutInfo> deserializeShortcuts(List<Map<String, String>> shortcuts) {
-    List<ShortcutInfo> shortcutInfos = new ArrayList<>();
-    Context context = registrar.context();
-    for (Map<String, String> shortcut : shortcuts) {
-      String icon = shortcut.get("icon");
-      String type = shortcut.get("type");
-      String title = shortcut.get("localizedTitle");
-      ShortcutInfo.Builder shortcutBuilder = new ShortcutInfo.Builder(context, type);
-      if (icon != null) {
-        int resourceId =
-            context.getResources().getIdentifier(icon, "drawable", context.getPackageName());
-        if (resourceId > 0) {
-          shortcutBuilder.setIcon(Icon.createWithResource(context, resourceId));
-        }
-      }
-      shortcutBuilder.setLongLabel(title);
-      shortcutBuilder.setShortLabel(title);
-      Intent intent = new Intent(context, ShortcutHandlerActivity.class);
-      intent.setAction("plugins.flutter.io/quick_action");
-      intent.putExtra("type", type);
-      shortcutBuilder.setIntent(intent);
-      shortcutInfos.add(shortcutBuilder.build());
-    }
-    return shortcutInfos;
+  @Override
+  public void onDetachedFromEngine(FlutterPluginBinding binding) {
+    teardownChannel();
   }
 
-  /**
-   * Handle the shortcut and immediately closes the activity.
-   *
-   * <p>Needs to be invocable by Android system; hence it is public.
-   */
-  public static class ShortcutHandlerActivity extends Activity {
+  @Override
+  public void onAttachedToActivity(ActivityPluginBinding binding) {
+    handler.setActivity(binding.getActivity());
+  }
 
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-      super.onCreate(savedInstanceState);
-      // Get the Intent that started this activity and extract the string
-      Intent intent = getIntent();
-      String type = intent.getStringExtra("type");
-      if (channel != null) {
-        channel.invokeMethod("launch", type);
-      }
-      finish();
-    }
+  @Override
+  public void onDetachedFromActivity() {
+    handler.setActivity(null);
+  }
+
+  @Override
+  public void onReattachedToActivityForConfigChanges(ActivityPluginBinding binding) {
+    onAttachedToActivity(binding);
+  }
+
+  @Override
+  public void onDetachedFromActivityForConfigChanges() {
+    onDetachedFromActivity();
+  }
+
+  private void setupChannel(BinaryMessenger messenger, Context context, Activity activity) {
+    channel = new MethodChannel(messenger, CHANNEL_ID);
+    handler = new MethodCallHandlerImpl(context, activity);
+    channel.setMethodCallHandler(handler);
+  }
+
+  private void teardownChannel() {
+    channel.setMethodCallHandler(null);
+    channel = null;
+    handler = null;
   }
 }
