@@ -5,24 +5,27 @@
 package io.flutter.plugins.webviewflutter;
 
 import android.content.Context;
+import android.view.View;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.StandardMessageCodec;
 import io.flutter.plugin.platform.PlatformView;
 import io.flutter.plugin.platform.PlatformViewFactory;
 import java.util.Map;
 
-public class WebViewFactory extends PlatformViewFactory {
+public final class WebViewFactory extends PlatformViewFactory {
   private final BinaryMessenger messenger;
+  private final View containerView;
 
-  public WebViewFactory(BinaryMessenger messenger) {
+  WebViewFactory(BinaryMessenger messenger, View containerView) {
     super(StandardMessageCodec.INSTANCE);
     this.messenger = messenger;
+    this.containerView = containerView;
   }
 
   @SuppressWarnings("unchecked")
   @Override
   public PlatformView create(Context context, int id, Object args) {
     Map<String, Object> params = (Map<String, Object>) args;
-    return new FlutterWebView(context, messenger, id, params);
+    return new FlutterWebView(context, messenger, id, params, containerView);
   }
 }
