@@ -1,3 +1,7 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 #import "FLTCookieManager.h"
 
 @implementation FLTCookieManager {
@@ -22,28 +26,24 @@
 
 - (void)clearCookies:(FlutterResult)result {
   if (@available(iOS 9.0, *)) {
-    [self clearCookiesIos9AndLater:result];
+    NSSet<NSString *> *websiteDataTypes = [NSSet setWithObject:WKWebsiteDataTypeCookies];
+    WKWebsiteDataStore *dataStore = [WKWebsiteDataStore defaultDataStore];
+
+    void (^deleteAndNotify)(NSArray<WKWebsiteDataRecord *> *) =
+        ^(NSArray<WKWebsiteDataRecord *> *cookies) {
+          BOOL hasCookies = cookies.count > 0;
+          [dataStore removeDataOfTypes:websiteDataTypes
+                        forDataRecords:cookies
+                     completionHandler:^{
+                       result(@(hasCookies));
+                     }];
+        };
+
+    [dataStore fetchDataRecordsOfTypes:websiteDataTypes completionHandler:deleteAndNotify];
   } else {
     // support for iOS8 tracked in https://github.com/flutter/flutter/issues/27624.
     NSLog(@"Clearing cookies is not supported for Flutter WebViews prior to iOS 9.");
   }
-}
-
-- (void)clearCookiesIos9AndLater:(FlutterResult)result {
-  NSSet *websiteDataTypes = [NSSet setWithArray:@[ WKWebsiteDataTypeCookies ]];
-  WKWebsiteDataStore *dataStore = [WKWebsiteDataStore defaultDataStore];
-
-  void (^deleteAndNotify)(NSArray<WKWebsiteDataRecord *> *) =
-      ^(NSArray<WKWebsiteDataRecord *> *cookies) {
-        BOOL hasCookies = cookies.count > 0;
-        [dataStore removeDataOfTypes:websiteDataTypes
-                      forDataRecords:cookies
-                   completionHandler:^{
-                     result(@(hasCookies));
-                   }];
-      };
-
-  [dataStore fetchDataRecordsOfTypes:websiteDataTypes completionHandler:deleteAndNotify];
 }
 
 @end
